requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
supabase==2.0.2
websocket-client==1.6.4
python-dotenv==1.0.0
redis==5.0.4
grpcio~=1.54.0
flasgger==0.9.7.1
protobuf~=3.20.0 